+++ src/Admin/StopWordAdmin.php
@@ -5,17 +5,18 @@
 
 namespace Zicht\Bundle\SolrBundle\Admin;
 
-use Sonata\AdminBundle\Admin\Admin;
+use Sonata\AdminBundle\Admin\AbstractAdmin;
 use Sonata\AdminBundle\Datagrid\DatagridMapper;
 use Sonata\AdminBundle\Datagrid\ListMapper;
 use Sonata\AdminBundle\Form\FormMapper;
 use Sonata\AdminBundle\Route\RouteCollection;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Zicht\Bundle\SolrBundle\Entity\StopWord;
 
 /**
- * Class StopWordAdmin
+ * @extends AbstractAdmin<StopWord>
  */
-class StopWordAdmin extends Admin
+class StopWordAdmin extends AbstractAdmin
 {
     /**
      * {@inheritDoc}